refactor(session-broker): create internal broker namespace
