+++ src/session-broker/daemonState.registration.test.ts
