+++ src/session-broker/daemonState.test.ts
