+++ src/session-broker/daemonState.ts
