+++ src/session-broker/sessionTerminalMetadata.ts
