+++ src/ui/App.tsx
@@ -7,8 +7,8 @@ import { useRenderer, useTerminalDimensions } from "@opentui/react";
 import { Suspense, lazy, useCallback, useEffect, useMemo, useState, useRef } from "react";
 import type { AppBootstrap, CliInput, LayoutMode } from "../core/types";
 import { canReloadInput, computeWatchSignature } from "../core/watch";
-import { HunkHostClient } from "../daemon/client";
-import type { ReloadedSessionResult } from "../daemon/types";
+import { HunkHostClient } from "../session-broker/client";
+import type { ReloadedSessionResult } from "../session-broker/types";
 import { MenuBar } from "./components/chrome/MenuBar";
 import { StatusBar } from "./components/chrome/StatusBar";
 import { DiffPane } from "./components/panes/DiffPane";
